refactor(canvas): break CanvasEditorScreen into sub-components
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
